🎨 migrated stores to typescript
