+++ client/index.d.ts
@@ -1,0 +1,13 @@
+import { accessorType } from '~/store';
+
+declare module 'vue/types/vue' {
+  interface Vue {
+    $accessor: typeof accessorType;
+  }
+}
+
+declare module '@nuxt/types' {
+  interface NuxtAppOptions {
+    $accessor: typeof accessorType;
+  }
+}
