--- client/store/environment.js
@@ -1,8 +1,0 @@
-export const state = () => ({
-  env: {}
-});
-export const getters = {
-  apiUrl(state) {
-    return state.env.apiUrl;
-  }
-};
