--- client/store/index.js
@@ -1,18 +1,0 @@
-export const actions = {
-  nuxtServerInit({ rootState, dispatch, getters }) {
-    if (process.server) {
-      rootState.environment.env = {
-        apiUrl: process.env.VIEWTUBE_API_URL,
-        vapidKey: process.env.VIEWTUBE_PUBLIC_VAPID,
-        nodeEnv: process.env.NODE_ENV,
-        host: process.env.HOST || 'localhost',
-        port: process.env.PORT || 8066,
-        baseUrl: process.env.BASE_URL || 'http://localhost:8066'
-      };
-      dispatch('user/getUser');
-      if (getters['instances/instances'].length === 0) {
-        dispatch('instances/fetchInstances');
-      }
-    }
-  }
-};
