--- client/store/videoPlayer.js
@@ -1,12 +1,0 @@
-export const state = () => ({
-  qualitySources: null
-});
-
-export const getters = {
-  resolutions: state => state.qualitySources
-};
-
-export const mutations = {
-  setResolutions: (state, qualitySources) =>
-    (state.resolutions = qualitySources)
-};
