+++ package.json
@@ -85,6 +85,7 @@
     "node-sass": "^4.14.1",
     "nuxt": "^2.14.6",
     "nuxt-start": "^2.14.6",
+    "nuxt-typed-vuex": "^0.1.22",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
@@ -102,6 +103,7 @@
     "universal-cookie": "^4.0.4",
     "vue": "^2.6.12",
     "vue-material-design-icons": "^4.9.0",
+    "vue-property-decorator": "^9.0.0",
     "vuex-persistedstate": "^3.1.0",
     "web-push": "^3.4.4",
     "x2js": "^3.4.0",
